requirements.txt
google-cloud-storage>=2.9.0
python-dotenv>=1.0.0
pandas>=2.0.0
pyarrow>=12.0.0
fastparquet>=2023.4.0
pyyaml>=6.0.0
# Dependências para conexão com Cloud SQL
sqlalchemy>=2.0.0
psycopg2-binary>=2.9.0  # Para PostgreSQL
pymysql>=1.0.0         # Para MySQL
cryptography>=41.0.0   # Para conexões seguras
google-cloud-secret-manager>=2.16.0  # Para gerenciamento de segredos
